+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 2022-01-24 - Edu
+### Added
+- **processes/process_redirections** Add redirections to process before executing
+
 ### Changed
 - **forks/close_quotes.c** Dinamyc CTRL+D error single double quotes open
 - rename t_redir as t_redirection